# contracts/cloud-api/v1/identity.openapi.yaml
openapi: 3.0.3
info:
  title: WT Media Cloud Identity API
  version: 2026.07.22.1
paths:
  /api/v1/auth/login:
    post:
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/LoginRequest'}
      responses:
        '200':
          description: Authenticated; sets the HttpOnly wt_media_session cookie.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/UserResponse'}
        '401': {$ref: '#/components/responses/AuthenticationRequired'}
  /api/v1/auth/me:
    get:
      operationId: getCurrentUser
      security: [{sessionCookie: []}]
      responses:
        '200':
          description: Current user.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/UserResponse'}
        '401': {$ref: '#/components/responses/AuthenticationRequired'}
  /api/v1/auth/logout:
    post:
      operationId: logout
      security: [{sessionCookie: []}]
      responses:
        '200': {description: Session invalidated.}
        '401': {$ref: '#/components/responses/AuthenticationRequired'}
  /api/v1/auth/change-password:
    post:
      operationId: changeOwnPassword
      security: [{sessionCookie: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ChangePasswordRequest'}
      responses:
        '200': {description: Password changed and existing session invalidated.}
        '401': {$ref: '#/components/responses/AuthenticationRequired'}
  /api/v1/users:
    get:
      operationId: listUsers
      security: [{sessionCookie: []}]
      parameters:
        - {in: query, name: uid, schema: {type: integer, format: int64}}
        - {in: query, name: username, schema: {type: string}}
        - {in: query, name: role, schema: {$ref: '#/components/schemas/Role'}}
        - {in: query, name: team_id, schema: {type: integer, format: int64}}
        - {in: query, name: status, schema: {type: string, enum: [enabled, disabled]}}
        - {in: query, name: game_id, schema: {type: string}}
      responses:
        '200':
          description: Filtered user list; never contains password material.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/UserListResponse'}
        '403': {$ref: '#/components/responses/Forbidden'}
    post:
      operationId: createUser
      security: [{sessionCookie: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/CreateUserRequest'}
      responses:
        '201':
          description: User created; the submitted initial password is returned only in this response.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/UserCreatedResponse'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '409': {description: Username already exists.}
  /api/v1/users/{user_id}:
    patch:
      operationId: updateUserAccess
      security: [{sessionCookie: []}]
      parameters:
        - in: path
          name: user_id
          required: true
          schema: {type: integer, format: int64}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/UpdateUserRequest'}
      responses:
        '200':
          description: User access updated.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/UserResponse'}
        '403': {$ref: '#/components/responses/Forbidden'}
    delete:
      operationId: deleteUser
      security: [{sessionCookie: []}]
      parameters:
        - in: path
          name: user_id
          required: true
          schema: {type: integer, format: int64}
      responses:
        '204': {description: Unreferenced user deleted.}
        '403': {$ref: '#/components/responses/Forbidden'}
        '409': {description: User still has sessions, audit, or business references.}
  /api/v1/users/{user_id}/reset-password:
    post:
      operationId: resetUserPassword
      security: [{sessionCookie: []}]
      parameters:
        - in: path
          name: user_id
          required: true
          schema: {type: integer, format: int64}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ResetPasswordRequest'}
      responses:
        '200':
          description: Password reset and sessions invalidated; plaintext returned only in this response.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/PasswordResetResponse'}
        '403': {$ref: '#/components/responses/Forbidden'}
  /api/v1/operation-teams:
    get:
      operationId: listOperationTeams
      security: [{sessionCookie: []}]
      responses:
        '200':
          description: Flat operation-team list.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/TeamListResponse'}
        '403': {$ref: '#/components/responses/Forbidden'}
    post:
      operationId: createOperationTeam
      security: [{sessionCookie: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/TeamRequest'}
      responses:
        '201':
          description: Team created.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/TeamResponse'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '409': {description: Team name already exists.}
  /api/v1/operation-teams/{team_id}:
    parameters:
      - in: path
        name: team_id
        required: true
        schema: {type: integer, format: int64}
    patch:
      operationId: renameOperationTeam
      security: [{sessionCookie: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/TeamRequest'}
      responses:
        '200':
          description: Team renamed.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/TeamResponse'}
        '403': {$ref: '#/components/responses/Forbidden'}
    delete:
      operationId: deleteOperationTeam
      security: [{sessionCookie: []}]
      responses:
        '204': {description: Unreferenced team deleted.}
        '403': {$ref: '#/components/responses/Forbidden'}
        '409': {description: Team is referenced by users or historical business objects.}
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: wt_media_session
  responses:
    AuthenticationRequired:
      description: Missing, invalid, or superseded session.
    Forbidden:
      description: Current role cannot perform the operation.
  schemas:
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username: {type: string, maxLength: 64}
        password: {type: string, format: password, minLength: 12}
    CreateUserRequest:
      type: object
      required: [username, password, role, game_ids]
      properties:
        username: {type: string, maxLength: 64}
        password: {type: string, format: password, minLength: 12}
        role: {$ref: '#/components/schemas/Role'}
        team_id: {type: integer, format: int64, nullable: true}
        game_ids: {type: array, items: {type: string}}
    UpdateUserRequest:
      type: object
      required: [role, game_ids]
      properties:
        role: {$ref: '#/components/schemas/Role'}
        team_id: {type: integer, format: int64, nullable: true}
        status: {type: string, enum: [enabled, disabled]}
        game_ids: {type: array, items: {type: string}}
    ChangePasswordRequest:
      type: object
      required: [current_password, new_password]
      properties:
        current_password: {type: string, format: password}
        new_password: {type: string, format: password, minLength: 12}
    ResetPasswordRequest:
      type: object
      required: [new_password]
      properties:
        new_password: {type: string, format: password, minLength: 12}
    PublicUser:
      type: object
      required: [id, username, role, status, team_id, team_name, game_ids]
      properties:
        id: {type: integer, format: int64}
        username: {type: string}
        role: {$ref: '#/components/schemas/Role'}
        status: {type: string, enum: [enabled, disabled]}
        team_id: {type: integer, format: int64, nullable: true}
        team_name: {type: string}
        game_ids: {type: array, items: {type: string}}
    Role:
      type: string
      enum: [admin, senior_operator, operator]
    OperationTeam:
      type: object
      required: [id, name, created_at, updated_at]
      properties:
        id: {type: integer, format: int64}
        name: {type: string, maxLength: 64}
        created_at: {type: string, format: date-time}
        updated_at: {type: string, format: date-time}
    TeamRequest:
      type: object
      required: [name]
      properties:
        name: {type: string, minLength: 1, maxLength: 64}
    UserResponse:
      type: object
      required: [data]
      properties:
        data: {$ref: '#/components/schemas/PublicUser'}
    UserCreatedResponse:
      type: object
      required: [data]
      properties:
        data:
          type: object
          required: [user, one_time_password]
          properties:
            user: {$ref: '#/components/schemas/PublicUser'}
            one_time_password: {type: string, writeOnly: true}
    PasswordResetResponse:
      type: object
      required: [data]
      properties:
        data:
          type: object
          required: [one_time_password]
          properties:
            one_time_password: {type: string, writeOnly: true}
    UserListResponse:
      type: object
      required: [data]
      properties:
        data: {type: array, items: {$ref: '#/components/schemas/PublicUser'}}
    TeamResponse:
      type: object
      required: [data]
      properties:
        data: {$ref: '#/components/schemas/OperationTeam'}
    TeamListResponse:
      type: object
      required: [data]
      properties:
        data: {type: array, items: {$ref: '#/components/schemas/OperationTeam'}}